+++ Lib/gftools/push.py
@@ -1,10 +1,303 @@
-"""Helper functions for creating/checking the server push files."""
-import requests
 from pathlib import Path
 from dataclasses import dataclass
+import os
 from gftools.utils import read_proto
 import gftools.fonts_public_pb2 as fonts_pb2
-from github import Github
+import requests
+from enum import Enum
+
+
+class PushCategory(Enum):
+    NEW = "New"
+    UPGRADE = "Upgrade"
+    OTHER = "Other"
+    DESIGNER_PROFILE = "Designer profile"
+    AXIS_REGISTRY = "Axis Registry"
+    KNOWLEDGE = "Knowledge"
+    METADATA = "Metadata / Description / License"
+    SAMPLE_TEXTS = "Sample texts"
+    BLOCKED = "Blocked"
+
+    def values():
+        return [i.value for i in PushCategory]
+
+
+FAMILY_FILE_SUFFIXES = frozenset([".ttf", ".otf", ".html", ".pb", ".txt"])
+
+
+GOOGLE_FONTS_TRAFFIC_JAM_QUERY = """
+{
+  organization(login: "google") {
+    projectV2(number: 74) {
+      id
+      title
+      items(last: 40) {
+        nodes {
+          id
+          status: fieldValueByName(name: "Status") {
+            ... on ProjectV2ItemFieldSingleSelectValue {
+              name
+            }
+          }
+          type
+          content {
+            ... on PullRequest {
+              id
+              files(first: 10) {
+                nodes {
+                  path
+                }
+              }
+              url
+              labels(first: 10) {
+                nodes {
+                  name
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
+"""
+
+
+@dataclass
+class PushItem:
+    path: Path
+    category: str
+    status: str
+    url: str
+
+    def __hash__(self):
+        return hash((self.path, self.category, self.status, self.url))
+
+    def exists(self):
+        return self.path.exists
+
+    def is_family(self):
+        return any(
+            t
+            for t in ("ofl", "apache", "ufl")
+            if t in self.path.parts
+            if "article" not in str(self.path)
+        )
+
+    def family_name(self):
+        assert self.is_family()
+        metadata_file = self.path / "METADATA.pb"
+        assert metadata_file.exists(), f"no metadata for {self.path}"
+        return read_proto(metadata_file, fonts_pb2.FamilyProto()).name
+
+    def to_json(self):
+        return {
+            "path": str(self.path),
+            "type": self.category.value,
+            "status": self.status,
+            "url": self.url,
+        }
+
+
+class PushItems(list):
+    def add(self, item):
+        # noto font projects projects often contain an article/ dir, we remove this
+        if "article" in item.path.parts:
+            item.path = item.path.parent
+
+        # for font families, we only want the dir e.g ofl/mavenpro/MavenPro[wght].ttf --> ofl/mavenpro
+        elif (
+            any(d in item.path.parts for d in ("ofl", "ufl", "apache"))
+            and item.path.suffix in FAMILY_FILE_SUFFIXES
+        ):
+            item.path = item.path.parent
+
+        # for lang and axisregistry .textproto files, we need a transformed path
+        elif (
+            any(d in item.path.parts for d in ("lang", "axisregistry"))
+            and item.path.suffix == ".textproto"
+        ):
+            item.path = repo_path_to_google_path(item.path)
+
+        # don't include any axisreg or lang file which don't end in textproto
+        elif (
+            any(d in item.path.parts for d in ("lang", "axisregistry"))
+            and item.path.suffix != ".textproto"
+        ):
+            return
+
+        # Skip if path if it's a parent dir e.g ofl/ apache/ axisregistry/
+        if len(item.path.parts) == 1:
+            return
+
+        # Skip if path is a parent of an existing path
+        if any(str(item.path) in str(p.path) for p in self):
+            return
+
+        # Pop any push items which are a child of the item's path
+        to_pop = None
+        for idx, i in enumerate(self):
+            if str(i.path) in str(item.path):
+                to_pop = idx
+                break
+        if to_pop:
+            self.pop(to_pop)
+        self.append(item)
+
+    def missing_paths(self):
+        res = []
+        for item in self:
+            path = item.path
+            if any(p in ("lang", "axisregistry") for p in path.parts):
+                path = google_path_to_repo_path(path)
+            if not path.exists():
+                res.append(path)
+        return res
+
+    def to_server_file(self, fp):
+        from collections import defaultdict
+
+        bins = defaultdict(set)
+        for item in self:
+            bins[item.category.value].add(item)
+
+        res = []
+        for tag in PushCategory.values():
+            if tag not in bins:
+                continue
+            res.append(f"# {tag}")
+            for item in sorted(bins[tag], key=lambda k: k.path):
+                res.append(f"{item.path} # {item.url}")
+            res.append("")
+        if isinstance(fp, str):
+            doc = open(fp, "w")
+        else:
+            doc = fp
+        doc.write("\n".join(res))
+
+    @classmethod
+    def from_server_file(cls, fp, status):
+        if isinstance(fp, (str, Path)):
+            doc = open(fp)
+        else:
+            doc = fp
+        results = cls()
+
+        lines = doc.read().split("\n")
+        category = "Unknown"
+        for line in lines:
+            if not line:
+                continue
+            if line.startswith("#"):
+                category = line[1:].strip()
+            elif "#" in line:
+                path, url = line.split("#")
+                item = PushItem(Path(path.strip()), category, status, url.strip())
+                results.add(item)
+            # some paths may not contain a PR, still add them
+            else:
+                item = PushItem(Path(line.strip()), category, status, "")
+                results.add(item)
+        return results
+
+    @classmethod
+    def from_traffic_jam(cls):
+        from gftools.gfgithub import GitHubClient
+
+        g = GitHubClient("google", "fonts")
+        data = g._run_graphql(GOOGLE_FONTS_TRAFFIC_JAM_QUERY, {})
+
+        board_items = data["data"]["organization"]["projectV2"]["items"]["nodes"]
+        results = cls()
+        for item in board_items:
+            status = item.get("status", {}).get("name", None)
+
+            if "labels" not in item["content"]:
+                print("PR missing labels. Skipping")
+            labels = [i["name"] for i in item["content"]["labels"]["nodes"]]
+
+            files = [Path(i["path"]) for i in item["content"]["files"]["nodes"]]
+            url = item["content"]["url"]
+
+            # get pr state
+            if "-- blocked" in labels:
+                cat = PushCategory.BLOCKED
+
+            if "I Font Upgrade" in labels or "I Small Fix" in labels:
+                cat = PushCategory.UPGRADE
+            elif "I New Font" in labels:
+                cat = PushCategory.NEW
+            elif "I Description/Metadata/OFL" in labels:
+                cat = PushCategory.METADATA
+            elif "I Designer profile" in labels:
+                cat = PushCategory.DESIGNER_PROFILE
+            elif "I Knowledge" in labels:
+                cat = PushCategory.KNOWLEDGE
+            elif "I Axis Registry" in labels:
+                cat = PushCategory.AXIS_REGISTRY
+            elif "I Lang" in labels:
+                cat = PushCategory.SAMPLE_TEXTS
+            else:
+                cat = PushCategory.OTHER
+
+            for f in files:
+                results.add(PushItem(Path(f), cat, status, url))
+        return results
+
+
+# The internal Google fonts team store the axisregistry and lang directories
+# in a different location. The two functions below tranform paths to
+# whichever representation you need.
+def repo_path_to_google_path(fp):
+    """lang/Lib/gflanguages/data/languages/.*.textproto --> lang/languages/.*.textproto"""
+    # we rename lang paths due to: https://github.com/google/fonts/pull/4679
+    if "gflanguages" in fp.parts:
+        return Path("lang") / fp.relative_to("lang/Lib/gflanguages/data")
+    # https://github.com/google/fonts/pull/5147
+    elif "axisregistry" in fp.parts:
+        return Path("axisregistry") / fp.name
+    return fp
+
+
+def google_path_to_repo_path(fp):
+    """lang/languages/.*.textproto --> lang/Lib/gflanguages/data/languages/.*.textproto"""
+    if "lang" in fp.parts:
+        return Path("lang/Lib/gflanguages/data/") / fp.relative_to("lang")
+    elif "axisregistry" in fp.parts:
+        return fp.parent / "Lib" / "axisregistry" / "data" / fp.name
+    return fp
+
+
+def lint_server_files(fp):
+    template = "{}: Following paths are not valid:\n{}\n\n"
+    footnote = (
+        "lang and axisregistry dir paths need to be transformed.\n"
+        "See https://github.com/googlefonts/gftools/issues/603"
+    )
+
+    prod_path = fp / "to_production.txt"
+    production_file = PushItems.from_server_file(prod_path, "In Sandbox")
+    prod_missing = "\n".join(map(str, production_file.missing_paths()))
+    prod_msg = template.format("to_production.txt", prod_missing)
+
+    sandbox_path = fp / "to_sandbox.txt"
+    sandbox_file = PushItems.from_server_file(sandbox_path, "In Dev / PR Merged")
+    sandbox_missing = "\n".join(map(str, sandbox_file.missing_paths()))
+    sandbox_msg = template.format("to_sandbox.txt", sandbox_missing)
+
+    if prod_missing and sandbox_missing:
+        raise ValueError(prod_msg + sandbox_msg + footnote)
+    elif prod_missing:
+        raise ValueError(prod_msg + footnote)
+    elif sandbox_missing:
+        raise ValueError(sandbox_msg + footnote)
+    else:
+        print("Server files have valid paths")
+
+
+# TODO refactor below server code once backend team have implemented a
+# tracking field to help us track prs through the servers
 
 
 SANDBOX_URL = "https://fonts.sandbox.google.com/metadata/fonts"
@@ -20,46 +313,6 @@ Existing families, last pushed:
 """
 
 
-@dataclass
-class PushItem:
-    path: Path
-    raw: str
-    type: str
-
-    def to_json(self):
-        return {"path": str(self.path), "type": self.type, "raw": self.raw}
-
-
-def parse_server_file(fp):
-    results = []
-    with open(fp) as doc:
-        lines = doc.read().split("\n")
-        category = "Unknown"
-        for line in lines:
-            if not line:
-                continue
-            if line.startswith("#"):
-                category = line[1:].strip()
-            elif "#" in line:
-                path = line.split("#")[0].strip()
-                item = PushItem(Path(path), line, category)
-                results.append(item)
-            else:
-                item = PushItem(Path(line), line, category)
-                results.append(item)
-    return results
-
-
-def is_family_dir(path):
-    return any(t for t in ("ofl", "apache", "ufl") if t in path.parts if "article" not in str(path))
-
-
-def family_dir_name(path):
-    metadata_file = path / "METADATA.pb"
-    assert metadata_file.exists(), f"no metadata for {path}"
-    return read_proto(metadata_file, fonts_pb2.FamilyProto()).name
-
-
 def gf_server_metadata(url):
     """Get family json data from a Google Fonts metadata url"""
     # can't do requests.get("url").json() since request text starts with ")]}'"
@@ -68,9 +321,9 @@ def gf_server_metadata(url):
 
 
 def server_push_status(fp, url):
-    dirs = [fp.parent / p.path for p in parse_server_file(fp)]
-    family_dirs = [d for d in dirs if is_family_dir(d)]
-    family_names = [family_dir_name(d) for d in family_dirs]
+    family_names = [
+        i.family_name() for i in PushItems.from_server_file(fp, "") if i.is_family()
+    ]
 
     gf_meta = gf_server_metadata(url)
 
@@ -97,66 +350,3 @@ def push_report(fp):
 
     sandbox_path = fp / "to_sandbox.txt"
     server_push_report("Sandbox", sandbox_path, SANDBOX_URL)
-
-
-# The internal Google fonts team store the axisregistry and lang directories
-# in a different location. The two functions below tranform paths to
-# whichever representation you need.
-def repo_path_to_google_path(fp):
-    """lang/Lib/gflanguages/data/languages/.*.textproto --> lang/languages/.*.textproto"""
-    # we rename lang paths due to: https://github.com/google/fonts/pull/4679
-    if "gflanguages" in fp.parts:
-        return Path("lang") / fp.relative_to("lang/Lib/gflanguages/data")
-    # https://github.com/google/fonts/pull/5147
-    elif "axisregistry" in fp.parts:
-        return Path("axisregistry") / fp.name
-    else:
-        raise ValueError(f"No transform found for path {fp}")
-
-
-def google_path_to_repo_path(fp):
-    """lang/languages/.*.textproto --> lang/Lib/gflanguages/data/languages/.*.textproto"""
-    if "lang" in fp.parts:
-        return Path("lang/Lib/gflanguages/data/") / fp.relative_to("lang")
-    elif "axisregistry" in fp.parts:
-        return fp.parent / "Lib" / "axisregistry" / "data" / fp.name
-    else:
-        raise ValueError(f"No transform found for path {fp}")
-
-
-def missing_paths(fp):
-    paths = [fp.parent / p.path for p in parse_server_file(fp)]
-    font_paths = [p for p in paths if any(d in p.parts for d in ("ofl", "ufl", "apache"))]
-    lang_paths = [p for p in paths if "lang" in p.parts]
-    axis_paths = [p for p in paths if "axisregistry" in p.parts]
-    misc_paths = [p for p in paths if p not in font_paths+lang_paths+axis_paths]
-
-    missing_paths = [p for p in misc_paths+font_paths if not p.exists()]
-    missing_lang_files = [p for p in lang_paths if not google_path_to_repo_path(p).exists()]
-    missing_axis_files = [p for p in axis_paths if not google_path_to_repo_path(p).exists()]
-    return missing_paths + missing_lang_files + missing_axis_files
-
-
-def lint_server_files(fp):
-    template = "{}: Following paths are not valid:\n{}\n\n"
-    footnote = (
-        "lang and axisregistry dir paths need to be transformed.\n"
-        "See https://github.com/googlefonts/gftools/issues/603"
-    )
-
-    prod_path = fp / "to_production.txt"
-    prod_missing = "\n".join(map(str, missing_paths(prod_path)))
-    prod_msg = template.format("to_production.txt", prod_missing)
-
-    sandbox_path = fp / "to_sandbox.txt"
-    sandbox_missing = "\n".join(map(str, missing_paths(sandbox_path)))
-    sandbox_msg = template.format("to_sandbox.txt", sandbox_missing)
-
-    if prod_missing and sandbox_missing:
-        raise ValueError(prod_msg + sandbox_msg + footnote)
-    elif prod_missing:
-        raise ValueError(prod_msg + footnote)
-    elif sandbox_missing:
-        raise ValueError(sandbox_msg + footnote)
-    else:
-        print("Server files have valid paths")
